Use key file if supplied in place of password
This is a hack to add rudimentary support for connecting to remote
hosts that only support key pair authentication. If a path to a public
key file is passed in in place of a password, attempt to authenticate
to the remote host using the key via the PublicKeys authentication
method.